refactor(call): extract _CallSession extension from CallBloc (#1017)
Move all per-call WebRTC state machine handlers (~690 lines) into a
dedicated part file (call_session.dart) as a private extension.

Includes:
- _onCallPerformEvent dispatcher and all 8 __onCallPerformEvent* handlers
  (started, answered, ended, setHeld, setMuted, sentDTMF, audioDeviceSet, audioDevicesUpdate)
- _onPeerConnectionEvent dispatcher and all 7 __onPeerConnectionEvent* handlers
  (signalingState, connectionState, iceGathering, iceConnection, iceCandidate, streamAdded, streamRemoved)
- _createPeerConnection, _handleRenegotiationNeeded, _checkSenderResult helpers
- _addToRecents, _playRingbackSound, _stopRingbackSound
- _assingUserActiveCalls, _assingNumberPresence, syncPresenceSettings

on<> registrations in CallBloc use lambda wrappers to call extension methods.
